refactor: split src/construct.c into src/construct/ modules (#95)
construct.c held construction end to end: the entry points, the
argument binding, the mutable-default copying, and set_field. This
splits it along those seams, moving the code verbatim:

  src/construct/construct.c  Struct_vectorcall, Struct_new, post-init
  src/construct/binding.c    argument binding, slot writing, find_field
  src/construct/defaults.c   mutable-default copying, set_field
  src/construct/construct.h  the public API, the field_lookup type, and
                             the cross-file declarations

The five public functions keep their old construct.h signatures, and
the six helpers that cross a boundary are declared there too; the rest
stay static in their file. The header includes types.h rather than
forward-declaring StructType, because on the pre-#68 types.h the
struct is anonymous and a forward typedef would name a different type.
The TESTING include gains a `../` for the quoted-include rule.

No behaviour change: the matrix gate (3.10-3.15, free-threaded, the
in-file C tests, the analyzers) passes on the split as on the original.
