+++ build_config.py
@@ -11,7 +11,9 @@ BUILD: Final = BuildConfig(
         "src/salix.c",
         "src/annotations.c",
         "src/compare.c",
-        "src/construct.c",
+        "src/construct/construct.c",
+        "src/construct/binding.c",
+        "src/construct/defaults.c",
         "src/fields.c",
         "src/hash.c",
         "src/meta/meta.c",
