+++ src/construct/construct.c
@@ -1,0 +1,78 @@
+#include <Python.h>
+
+#include "construct.h"
+#include "../owned.h"
+#include "../result.h"
+
+static enum result run_post_init(StructType const * type, PyObject * self);
+
+PyObject * Struct_vectorcall(
+	PyObject * const struct_class,
+	PyObject * const * const arguments,
+	size_t const argument_count_and_flags,
+	PyObject * const keyword_names
+) {
+	StructType * const type = (StructType *) struct_class;
+	Py_ssize_t const positional_count = PyVectorcall_NARGS(argument_count_and_flags);
+
+	if (positional_count > type->struct_field_count) {
+		PyErr_Format(
+			PyExc_TypeError,
+			"%.200s() takes at most %zd positional arguments but %zd were given",
+			struct_type_name(type),
+			type->struct_field_count,
+			positional_count
+		);
+
+		return NULL;
+	}
+
+	PyTypeObject * const python_class = &type->heap_type.ht_type;
+
+	PY_MOVABLE(self, python_class->tp_alloc(python_class, 0));
+
+	if (self == NULL) {
+		return NULL;
+	}
+
+	bind_positional(type, self, arguments, positional_count);
+
+	if (
+		bind_keywords(type, self, arguments, positional_count, keyword_names) != RESULT_OK ||
+		fill_defaults(type, self, positional_count) != RESULT_OK ||
+		run_post_init(type, self) != RESULT_OK
+	) {
+		return NULL;
+	}
+
+	return py_move(&self);
+}
+
+PyObject * Struct_new(
+	PyTypeObject * const struct_class,
+	PyObject * const arguments,
+	PyObject * const keywords
+) {
+	PY_MOVABLE(self, struct_class->tp_alloc(struct_class, 0));
+
+	if (self == NULL) {
+		return NULL;
+	}
+
+	StructType const * const type = (StructType *) struct_class;
+
+	return (
+		fill_defaults(type, self, struct_required_count(type)) == RESULT_OK ? py_move(&self) :
+		NULL
+	);
+}
+
+static enum result run_post_init(StructType const * const type, PyObject * const self) {
+	if (type->struct_post_init == NULL) {
+		return RESULT_OK;
+	}
+
+	PY_OWNED(returned, PyObject_CallOneArg(type->struct_post_init, self));
+
+	return returned != NULL ? RESULT_OK : RESULT_ERROR;
+}
