+++ src/construct/defaults.c
@@ -1,0 +1,98 @@
+#include <Python.h>
+
+#include "construct.h"
+#include "../result.h"
+
+/* A list of statements rather than a static array of {type, constructor}: on
+ * Windows a `PyTypeObject` is imported from python3.dll, and the address of a
+ * dllimport symbol is not a compile-time constant, so the array version
+ * compiles everywhere except the platform half the wheels are cross-built for.
+ */
+typedef PyObject * (*default_copier)(PyObject * declared);
+
+static PyObject * copy_list(PyObject * const declared) {
+	return PyList_GetSlice(declared, 0, PyList_GET_SIZE(declared));
+}
+
+static default_copier copies_default(PyTypeObject const * const kind) {
+	if (kind == &PyList_Type) {
+		return copy_list;
+	}
+
+	if (kind == &PyDict_Type) {
+		return PyDict_Copy;
+	}
+
+	if (kind == &PySet_Type) {
+		return PySet_New;
+	}
+
+	if (kind == &PyByteArray_Type) {
+		return PyByteArray_FromObject;
+	}
+
+	return NULL;
+}
+
+bool struct_copies_default(PyTypeObject const * const kind) {
+	return copies_default(kind) != NULL;
+}
+
+PyObject * struct_default_copy(PyObject * const declared) {
+	default_copier const copy = copies_default(Py_TYPE(declared));
+
+	return copy != NULL ? copy(declared) : Py_NewRef(declared);
+}
+
+PyObject * Struct_set_field(PyObject * const module, PyObject * const arguments) {
+	PyObject * self = NULL;
+	PyObject * name = NULL;
+	PyObject * value = NULL;
+
+	if (!PyArg_UnpackTuple(arguments, "set_field", 3, 3, &self, &name, &value)) {
+		return NULL;
+	}
+
+	if (!is_struct(self)) {
+		PyErr_Format(
+			PyExc_TypeError,
+			"set_field() expects a struct, not %.200s",
+			Py_TYPE(self)->tp_name
+		);
+
+		return NULL;
+	}
+
+	if (!PyUnicode_Check(name)) {
+		PyErr_Format(
+			PyExc_TypeError,
+			"set_field() field name must be str, not %.200s",
+			Py_TYPE(name)->tp_name
+		);
+
+		return NULL;
+	}
+
+	StructType * const type = struct_type_of(self);
+	struct field_lookup const found = find_field(type, name);
+
+	switch (found.tag) {
+		case FIELD_LOOKUP_ERROR:
+			return NULL;
+		case FIELD_LOOKUP_MISSING:
+			PyErr_Format(
+				PyExc_AttributeError,
+				"%.200s has no field '%U'",
+				struct_type_name(type),
+				name
+			);
+
+			return NULL;
+		case FIELD_LOOKUP_FOUND:
+			if (write_slot(type, self, found.index, value) != RESULT_OK) {
+				return NULL;
+			}
+	}
+
+	Py_RETURN_NONE;
+}
